docs: move constants to model folder
